Move 64-bit word into primitives
