+++ library/Rattletrap/AttributeValue/QWord.hs
@@ -1,6 +1,6 @@
 module Rattletrap.AttributeValue.QWord where
 
-import Rattletrap.Word64
+import Rattletrap.Primitive.Word64
 
 import qualified Data.Binary.Bits.Get as BinaryBit
 import qualified Data.Binary.Bits.Put as BinaryBit
