+++ library/Rattletrap/Json.hs
@@ -33,7 +33,7 @@ import Rattletrap.Section
 import Rattletrap.Primitive.Text
 import Rattletrap.Primitive.Vector
 import Rattletrap.Primitive.Word32
-import Rattletrap.Word64
+import Rattletrap.Primitive.Word64
 import Rattletrap.Primitive.Word8
 
 import qualified Control.Monad as Monad
