+++ library/Rattletrap/Property.hs
@@ -2,7 +2,7 @@ module Rattletrap.Property where
 
 import Rattletrap.PropertyValue
 import Rattletrap.Primitive.Text
-import Rattletrap.Word64
+import Rattletrap.Primitive.Word64
 
 import qualified Data.Binary as Binary
